requirements.txt
fastapi
langchain
langchain-community
langchain-openai
openai
uvicorn
python-dotenv
chromadb
sentence-transformers

# 语音转文字相关依赖
numpy>=1.21.0
librosa>=0.9.0
soundfile>=0.10.0
openai-whisper>=20231117
pyaudio>=0.2.11 